Database functions for program 2 now functional atleast logically
Database functions for program 2 have been implemented and functions from program 1 is now updated. The program is multi-threaded now meaning that multiple clients can connect to server at the same time. Will be writing a new issue to GitHub repository in the next few days with a more complete breakdown of the update as well as what is left to do and check.